Actually lift the profile endpoints (the last commit shipped an orphan)
Post-work audit caught it: the four-cluster commit included api/user_profiles.py
but web_server still owned all 27 profile routes and never wired the blueprint -
the late-night git-checkout revert of the failed batch also threw away the
profiles carve and its wiring, while the half-extracted module file survived and
got committed as dead code (and the commit message counted it as lifted).

This is the real lift: 31 routes (profiles + per-profile credentials), the
service-connection helper block, session/limiter/page-id injection, getters for
the rebindable clients. Verified: url-map byte-identical on method+path with all
31 endpoints renamed to user_profiles.*, no duplicate rules, smokes green,
profile suites green.
